requirements.txt
asgiref==3.3.4
Django==3.2
django-cors-headers==3.7.0
django-cors-middleware==1.5.0
djangorestframework==3.12.4
gunicorn==20.1.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
whitenoise==5.2.0
